requirements.txt
behave
playwright
python-dotenv
from_root
allure-behave
